# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: xg2g API
  description: |
    API for xg2g (Enigma2 to Plex/Jellyfin Proxy).

    **Authentication**:
    All endpoints support both **Bearer Token** (`Authorization: Bearer <token>`) and **Cookie** (`xg2g_session=<token>`) authentication.
    Streaming and media requests specifically support unified authentication since v3.1.5.

    **License**: PolyForm Noncommercial License 1.0.0.
    Commercial usage is restricted. See repository documentation for details.
  version: 3.0.0
servers:
  - url: /api/v3
    description: Production API
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    DeviceAuthDeviceType:
      type: string
      enum: [android_phone, android_tablet, android_tv, browser, unknown]
    PairingStatus:
      type: string
      enum: [pending, approved, expired, consumed, revoked]
    PublishedEndpointKind:
      type: string
      enum: [public_https, local_https, local_http]
    PublishedEndpointTLSMode:
      type: string
      enum: [required, prohibited]
    PublishedEndpointSource:
      type: string
      enum: [config, env, operator]
...
    PublishedEndpoint:
      type: object
      additionalProperties: false
      required:
        [url, kind, priority, tlsMode, allowPairing, allowStreaming, allowWeb, allowNative, advertiseReason, source]
      properties:
        url:
          type: string
          format: uri
        kind:
          $ref: "#/components/schemas/PublishedEndpointKind"
        priority:
          type: integer
          format: int32
        tlsMode:
          $ref: "#/components/schemas/PublishedEndpointTLSMode"
        allowPairing:
          type: boolean
        allowStreaming:
          type: boolean
        allowWeb:
          type: boolean
        allowNative:
          type: boolean
        advertiseReason:
          type: string
        source:
          $ref: "#/components/schemas/PublishedEndpointSource"
...
    StartPairingRequest:
      type: object
      additionalProperties: false
      properties:
        deviceName:
          type: string
        deviceType:
          $ref: "#/components/schemas/DeviceAuthDeviceType"
        requestedPolicyProfile:
          type: string
    PairingSecretRequest:
      type: object
      additionalProperties: false
      required: [pairingSecret]
      properties:
        pairingSecret:
          type: string
    ApprovePairingRequest:
      type: object
      additionalProperties: false
      properties:
        ownerId:
          type: string
        approvedPolicyProfile:
          type: string
    StartPairingResponse:
      type: object
      additionalProperties: false
      required: [pairingId, pairingSecret, userCode, qrPayload, expiresAt]
      properties:
        pairingId:
          type: string
        pairingSecret:
          type: string
        userCode:
          type: string
        qrPayload:
          type: string
        expiresAt:
          type: string
          format: date-time
    PairingStatusResponse:
      type: object
      additionalProperties: false
      required: [pairingId, status, userCode, deviceName, deviceType, expiresAt]
      properties:
        pairingId:
          type: string
        status:
          $ref: "#/components/schemas/PairingStatus"
        userCode:
          type: string
        deviceName:
          type: string
        deviceType:
          $ref: "#/components/schemas/DeviceAuthDeviceType"
        requestedPolicyProfile:
          type: string
        approvedPolicyProfile:
          type: string
        expiresAt:
          type: string
          format: date-time
        approvedAt:
          type: string
          format: date-time
        consumedAt:
          type: string
          format: date-time
    ApprovePairingResponse:
      type: object
      additionalProperties: false
      required: [pairingId, status, ownerId, expiresAt]
      properties:
        pairingId:
          type: string
        status:
          $ref: "#/components/schemas/PairingStatus"
        ownerId:
          type: string
        approvedPolicyProfile:
          type: string
        approvedAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
    ExchangePairingResponse:
      type: object
      additionalProperties: false
      required:
        [
          pairingId,
          deviceId,
          deviceGrantId,
          deviceGrant,
          deviceGrantExpiresAt,
          accessSessionId,
          accessToken,
          accessTokenExpiresAt,
          policyVersion,
          scopes,
          endpoints,
        ]
      properties:
        pairingId:
          type: string
        deviceId:
          type: string
        deviceGrantId:
          type: string
        deviceGrant:
          type: string
        deviceGrantExpiresAt:
          type: string
          format: date-time
        accessSessionId:
          type: string
        accessToken:
          type: string
        accessTokenExpiresAt:
          type: string
          format: date-time
        policyVersion:
          type: string
        scopes:
          type: array
          items:
            type: string
        endpoints:
          type: array
          items:
            $ref: "#/components/schemas/PublishedEndpoint"
...
    ProblemCode:
      type: string
      description: Registry-backed public machine-readable short code for RFC7807 responses.
      enum:
        - ADD_FAILED
        - ADMISSION_ENGINE_DISABLED
        - ADMISSION_NO_TUNERS
        - ADMISSION_SESSIONS_FULL
        - ADMISSION_STATE_UNKNOWN
        - ADMISSION_TRANSCODES_FULL
        - ADMISSION_UNAVAILABLE
        - BOUQUET_NOT_FOUND
        - BREAKER_OPEN
        - CLAIM_MISMATCH
        - CLIENT_UNAVAILABLE
        - CONCURRENT_BUILDS_EXCEEDED
        - CONFLICT
        - DELETE_FAILED
        - DIFF_FAILED
        - DURATION_INVALID
        - DURATION_NEGATIVE
        - DURATION_OVERFLOW
        - ENGINE_ERROR
        - FILE_NOT_FOUND
        - FORBIDDEN
        - HTTPS_REQUIRED
        - INTERNAL_ERROR
        - INTERNAL_SERVER_ERROR
        - INVALID_CAPABILITIES
        - INVALID_ID
        - INVALID_INPUT
        - INVALID_PLAYLIST_PATH
        - INVALID_SESSION_ID
        - INVALID_TIME
        - INVALID_TOKEN
        - LIBRARY_ROOT_NOT_FOUND
        - LIBRARY_SCAN_RUNNING
        - METHOD_NOT_ALLOWED
        - NOT_FOUND
        - NOT_IMPLEMENTED
        - PANIC
        - PATH_TRAVERSAL
        - PREFLIGHT_BAD_GATEWAY
        - PREFLIGHT_FORBIDDEN
        - PREFLIGHT_INTERNAL
        - PREFLIGHT_NOT_FOUND
        - PREFLIGHT_TIMEOUT
        - PREFLIGHT_UNAUTHORIZED
        - PREFLIGHT_UNREACHABLE
        - PREPARING
        - PROVIDER_ERROR
        - RATE_LIMIT_EXCEEDED
        - READ_FAILED
        - RECEIVER_ERROR
        - RECEIVER_INCONSISTENT
        - RECEIVER_UNREACHABLE
        - RECORDING_NOT_FOUND
        - RECORDING_PREPARING
        - REFRESH_FAILED
        - REFRESH_IN_PROGRESS
        - REMOTE_PROBE_UNSUPPORTED
        - R_RECORDING_NOT_READY
        - SAVE_FAILED
        - SCAN_UNAVAILABLE
        - SECURITY_UNAVAILABLE
        - SERVICE_NOT_FOUND
        - SERVICE_UNAVAILABLE
        - SESSION.EXPIRED
        - SESSION.NOT_FOUND
        - SESSION.UPDATE_ERROR
        - SESSION_NOT_FOUND
        - STOP_FAILED
        - STORE_ERROR
        - TOKEN_AUD_MISMATCH
        - TOKEN_CAP_MISMATCH
        - TOKEN_ERROR
        - TOKEN_EXPIRED
        - TOKEN_INVALID_ALG
        - TOKEN_INVALID_SIG
        - TOKEN_ISS_MISMATCH
        - TOKEN_MALFORMED
        - TOKEN_MISSING
        - TOKEN_MISSING_CLAIM
        - TOKEN_MODE_MISMATCH
        - TOKEN_NOT_ACTIVE
        - TOKEN_SUB_MISMATCH
        - TOKEN_TTL_EXCEEDED
        - TRANSCODE_CANCELED
        - TRANSCODE_FAILED
        - TRANSCODE_STALLED
        - TRANSCODE_START_TIMEOUT
        - UNAUTHORIZED
        - UNAVAILABLE
        - UPDATE_FAILED
        - UPSTREAM_AUTH
        - UPSTREAM_EMPTY
        - UPSTREAM_ERROR
        - UPSTREAM_RESULT_FALSE
        - UPSTREAM_TIMEOUT
        - UPSTREAM_UNAVAILABLE
        - V3_UNAVAILABLE
        - capabilities_invalid
        - capabilities_missing
        - decision_ambiguous
        - invariant_violation
        - session_gone
...
    ProblemDetails:
      type: object
      required: [type, title, status, requestId]
      additionalProperties: false
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        requestId:
          type: string
          description: Correlation ID (UUID or prefixed string like req_abc123)
        code:
          $ref: "#/components/schemas/ProblemCode"
        detail: { type: string }
        instance: { type: string }
        fields:
          type: object
          additionalProperties: true
        conflicts:
          type: array
          items: { $ref: "#/components/schemas/TimerConflict" }
    Timer:
      type: object
      required: [timerId, serviceRef, begin, end, name, state]
      properties:
        timerId: { type: string }
        serviceRef: { type: string }
        begin: { type: integer, format: int64 }
        end: { type: integer, format: int64 }
        name: { type: string }
        description: { type: string }
        serviceName: { type: string }
        state:
          type: string
          enum: [scheduled, recording, completed, disabled, unknown]
        receiverState:
          type: object
          additionalProperties: true
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    TimerConflict:
      type: object
      required: [type, blockingTimer]
      properties:
        type:
          type: string
          enum: [overlap, duplicate, tuner_limit, unknown]
        blockingTimer: { $ref: "#/components/schemas/Timer" }
        overlapSeconds: { type: integer }
        message: { type: string }
...
    LogEntry:
      type: object
      properties:
        time:
          type: string
          format: date-time
        level:
          type: string
        message:
          type: string
        fields:
          type: object
          additionalProperties: true
...
security:
  - BearerAuth: [v3:read]
paths:
...
  /pairing/start:
    post:
      summary: Start device pairing enrollment
      operationId: StartPairing
      tags:
        - auth
      security: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartPairingRequest"
      responses:
        "201":
          description: Pairing enrollment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StartPairingResponse"
        "400":
          description: Invalid pairing request
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "503":
          description: Pairing state store unavailable
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /pairing/{pairingId}/status:
    post:
      summary: Read pairing enrollment status
      operationId: GetPairingStatus
      tags:
        - auth
      security: []
      parameters:
        - name: pairingId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PairingSecretRequest"
      responses:
        "200":
          description: Pairing enrollment status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PairingStatusResponse"
        "400":
          description: Invalid pairing request
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "403":
          description: Pairing secret mismatch
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          description: Pairing not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "410":
          description: Pairing has expired or was revoked
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "503":
          description: Pairing state store unavailable
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /pairing/{pairingId}/approve:
    post:
      summary: Approve a pending pairing enrollment
      operationId: ApprovePairing
      tags:
        - auth
      security:
        - BearerAuth: [v3:admin]
      parameters:
        - name: pairingId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovePairingRequest"
      responses:
        "200":
          description: Pairing enrollment approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovePairingResponse"
        "400":
          description: Invalid pairing approval request
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "401":
          description: Authentication required
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          description: Pairing not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "409":
          description: Pairing is still pending or otherwise cannot transition
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "410":
          description: Pairing has expired, been revoked, or was already exchanged
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "503":
          description: Pairing state store unavailable
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /pairing/{pairingId}/exchange:
    post:
      summary: Exchange an approved pairing for device credentials
      operationId: ExchangePairing
      tags:
        - auth
      security: []
      parameters:
        - name: pairingId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PairingSecretRequest"
      responses:
        "200":
          description: Pairing exchanged for device grant and access session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExchangePairingResponse"
        "400":
          description: Invalid pairing exchange request
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "403":
          description: Pairing secret mismatch
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          description: Pairing not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "409":
          description: Pairing is pending and not yet approved
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "410":
          description: Pairing has expired, been revoked, or was already exchanged
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "503":
          description: Pairing state store unavailable
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
...
  /logs:
    get:
      summary: Get recent logs
      security:
        - BearerAuth: [v3:admin]
      operationId: getLogs
      tags:
        - system
      parameters:
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
          description: Maximum number of most-recent log entries to return.
      responses:
        "200":
          description: Recent log entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LogEntry"
...